Make the licence more general to easy the adoption
The licence have a new format for people to fill in their information. Also there is a change to reduce the free of charge use permission to personal use. This makes more clear that all "freedom" rights are granted but the use needs a licence, which can be a free one if the owner is contacted or the software is used for personal use.

In this case personal use doesn't involve a "no monetary" license clause. This personal use involves a use of the software where only one person is using it, not a whole organization. This is ambiguous in the sense that inside a organization only one person uses at a time, and every use can be justified as "personal use", but if the software is used inside a organization as part of their technologic stack this use case would not be a good reason to not pay a use licence.